componentize Search functions
